Move `Relation<T>` from 'lib.rs' into separate file 'relation.rs'
